fix(glean): extract engagementDetails into new file
